\name{CCA-package}
\alias{CCA-package}
\alias{CCA}
\encoding{latin1}
\docType{package}
\title{Canonical correlation analysis}

\description{The package provides a set of functions that extend the
\code{cancor()} function with new numerical and graphical outputs. It
includes a regularized extension of the canonical correlation analysis
to deal with datasets with more variables than observations and enables
to handle with missing values.}

\author{Ignacio Gonzalez, Sebastien Dejean
  Maintainer: Sebastien Dejean <sebastien.dejean@math.univ-toulouse.fr>}


\keyword{multivariate}