\name{img.estim.regul}
\alias{img.estim.regul}
\encoding{latin1}
\title{Plot the cross-validation criterion}
\description{
This function provide a visualization of the values of the
cross-validation criterion  obtained on a grid defined in the function \code{estim.regul()}.}
\usage{
img.estim.regul(estim)
}

\arguments{
  \item{estim}{Object returned by \code{estim.regul()}.}
}
\details{

}


\author{Sébastien Déjean, Ignacio González}


\seealso{\code{\link{estim.regul}}}
\examples{
}
\keyword{dplot}