\name{CCA-package}
\alias{CCA-package}
\alias{CCA}
\encoding{latin1}
\docType{package}
\title{Canonical correlation analysis}

\description{The package provides a set of functions that extend the
\code{cancor()} function with new numerical and graphical outputs. It
includes a regularized extension of the canonical correlation analysis
to deal with datasets with more variables than observations and enables
to handle with missing values.}

\details{
\tabular{ll}{
Package: \tab CCA\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2007-02-02\cr
License: \tab None\cr
}

}
\author{Ignacio González, Sébastien Déjean

  Maintainer: Sébastien Déjean <sebastien.dejean@math.ups-tlse.fr>
}

\references{
www.lsp.ups-tlse.fr/CCA
}
\keyword{multivariate}